front_parcialmente_finalizado

+++ IceTalk_FrontEnd/site/folders/delete/delete.js
@@ -1,11 +1,16 @@
 document.getElementById("myForm").addEventListener("submit", function (event) {
     event.preventDefault(); // Evita o envio do formulário padrão
 
-    formData = document.getElementById("cpf").value; // Obtém os dados do formulário
-    console.log(formData)
+    var formData = document.getElementById("cpf").value; // Obtém os dados do formulário
+
+    // Verifica se o campo CPF não está vazio
+    if (!formData.trim()) {
+        pushNotify('error', "Campo vazio", "Por favor, insira um CPF válido");
+        return;
+    }
+
     var url = 'http://localhost:8080/api/funcionarios/' + formData;
     var textarea = document.getElementById("textarea");
-    console.log(url)
 
     fetch(url, {
         method: 'DELETE',
@@ -15,15 +20,25 @@ document.getElementById("myForm").addEventListener("submit", function (event) {
     })
     .then(response => {
         if (!response.ok) {
-            pushNotify('error', "Vazio", "escreva algo");
-            throw new Error('Erro ao enviar dados.');
-        } else {
-            pushNotify("success", "Sucesso", "Dados excluidos com sucesso");
+            pushNotify('error', "Erro", "Ocorreu um erro ao excluir os dados");
+            throw new Error('Erro ao excluir dados.');
         }
-        return response.json();
+        // Verifica se a resposta é um JSON válido
+        const contentType = response.headers.get('content-type');
+        if (contentType && contentType.includes('application/json')) {
+            return response.json();
+        } else {
+            // Se a resposta não for JSON, retorna um objeto vazio
+            return {};
+        }
+    })
+    .then(data => {
+        // Se necessário, faça algo com a resposta da API
+        pushNotify("success", "Sucesso", "Dados excluídos com sucesso");
     })
     .catch(error => {
         console.error('Erro:', error);
+        pushNotify('error', "Erro", "Ocorreu um erro ao excluir os dados");
     });
 });
